move package protections rspec support into exported part of code
