+++ lib/package_protections/rspec/application_fixture_helper.rb
